fix: Move router pipeline creation under router.
Add utils file and move support functions there instead of router.
